refactor: move state and connect to cmf
