+++ packages/cmf/src/cmfConnect.test.js
