+++ packages/containers/src/SidePanel/SidePanel.connect.js
@@ -1,4 +1,4 @@
-import cmfConnect from '../cmfConnect';
+import { cmfConnect } from 'react-cmf';
 import Container, { DEFAULT_STATE } from './SidePanel.container';
 
 export default cmfConnect({
